+++ include/config/config.h
@@ -106,6 +106,7 @@ namespace bongocat::config {
 
         // for keep old index when reload config
         int32_t keep_old_animation_index{0};
+        int32_t strict{0};
 
 
         // Make Config movable and copyable
@@ -151,7 +152,8 @@ namespace bongocat::config {
               idle_animation(other.idle_animation),
               input_fps(other.input_fps),
               randomize_index(other.randomize_index),
-              keep_old_animation_index(other.keep_old_animation_index)
+              keep_old_animation_index(other.keep_old_animation_index),
+              strict(other.strict)
         {
             output_name = other.output_name ? strdup(other.output_name) : nullptr;
             config_copy_keyboard_devices_from(*this, other);
@@ -194,6 +196,7 @@ namespace bongocat::config {
                 input_fps = other.input_fps;
                 randomize_index = other.randomize_index;
                 keep_old_animation_index = other.keep_old_animation_index;
+                strict = other.strict;
 
                 output_name = other.output_name ? strdup(other.output_name) : nullptr;
                 config_copy_keyboard_devices_from(*this, other);
@@ -236,7 +239,8 @@ namespace bongocat::config {
               idle_animation(other.idle_animation),
               input_fps(other.input_fps),
               randomize_index(other.randomize_index),
-              keep_old_animation_index(other.keep_old_animation_index)
+              keep_old_animation_index(other.keep_old_animation_index),
+              strict(other.strict)
         {
             for (int i = 0; i < num_keyboard_devices; ++i) {
                 keyboard_devices[i] = other.keyboard_devices[i];
@@ -285,6 +289,7 @@ namespace bongocat::config {
                 input_fps = other.input_fps;
                 randomize_index = other.randomize_index;
                 keep_old_animation_index = other.keep_old_animation_index;
+                strict = other.strict;
 
                 for (int i = 0; i < num_keyboard_devices; ++i) {
                     keyboard_devices[i] = other.keyboard_devices[i];
@@ -321,7 +326,8 @@ namespace bongocat::config {
 
     struct load_config_overwrite_parameters_t {
         const char* output_name{nullptr};
-        int32_t random_index{-1};
+        int32_t randomize_index{-1};
+        int32_t strict{-1};
     };
     [[nodiscard]] created_result_t<config_t> load(const char *config_file_path, load_config_overwrite_parameters_t overwrite_parameters);
     void reset(config_t& config);