+++ src/config/config.cpp
@@ -107,51 +107,61 @@ namespace bongocat::config {
     // CONFIGURATION VALIDATION MODULE
     // =============================================================================
 
-    static constexpr void config_clamp_int(int& value, int min, int max, [[maybe_unused]] const char *name) {
+    static constexpr int32_t config_clamp_int(int& value, int min, int max, [[maybe_unused]] const char *name) {
         if (value < min || value > max) {
             BONGOCAT_LOG_WARNING("%s %d out of range [%d-%d], clamping", name, value, min, max);
             value = (value < min) ? min : max;
+            return 1;
         }
+        return 0;
     }
 
-    static void config_validate_dimensions(config_t& config) {
-        config_clamp_int(config.cat_height, MIN_CAT_HEIGHT, MAX_CAT_HEIGHT, CAT_HEIGHT_KEY);
-        config_clamp_int(config.overlay_height, MIN_OVERLAY_HEIGHT, MAX_OVERLAY_HEIGHT, OVERLAY_HEIGHT_KEY);
+    static int32_t config_validate_dimensions(config_t& config) {
+        int32_t ret{0};
+        ret |= config_clamp_int(config.cat_height, MIN_CAT_HEIGHT, MAX_CAT_HEIGHT, CAT_HEIGHT_KEY);
+        ret |= config_clamp_int(config.overlay_height, MIN_OVERLAY_HEIGHT, MAX_OVERLAY_HEIGHT, OVERLAY_HEIGHT_KEY);
+        return ret;
     }
 
-    static void config_validate_timing(config_t& config) {
-        config_clamp_int(config.fps, MIN_FPS, MAX_FPS, FPS_KEY);
-        config_clamp_int(config.keypress_duration_ms, MIN_DURATION_MS, MAX_DURATION_MS, KEYPRESS_DURATION_KEY);
-        config_clamp_int(config.test_animation_duration_ms, 0, MAX_DURATION_MS, TEST_ANIMATION_DURATION_KEY);
-        config_clamp_int(config.animation_speed_ms, 0, MAX_DURATION_MS, TEST_ANIMATION_DURATION_KEY);
-        config_clamp_int(config.idle_sleep_timeout_sec, 0, MAX_TIMEOUT, IDLE_SLEEP_TIMEOUT_KEY);
-        config_clamp_int(config.input_fps, 0, MAX_FPS, INPUT_FPS_KEY);
+    static int32_t config_validate_timing(config_t& config) {
+        int32_t ret{0};
+        ret |= config_clamp_int(config.fps, MIN_FPS, MAX_FPS, FPS_KEY);
+        ret |= config_clamp_int(config.keypress_duration_ms, MIN_DURATION_MS, MAX_DURATION_MS, KEYPRESS_DURATION_KEY);
+        ret |= config_clamp_int(config.test_animation_duration_ms, 0, MAX_DURATION_MS, TEST_ANIMATION_DURATION_KEY);
+        ret |= config_clamp_int(config.animation_speed_ms, 0, MAX_DURATION_MS, TEST_ANIMATION_DURATION_KEY);
+        ret |= config_clamp_int(config.idle_sleep_timeout_sec, 0, MAX_TIMEOUT, IDLE_SLEEP_TIMEOUT_KEY);
+        ret |= config_clamp_int(config.input_fps, 0, MAX_FPS, INPUT_FPS_KEY);
 
         // Validate interval (0 is allowed to disable)
         if (config.test_animation_interval_sec < 0 || config.test_animation_interval_sec > MAX_INTERVAL_SEC) {
             BONGOCAT_LOG_WARNING("%s %d out of range [0-%dsec], clamping",
                                  TEST_ANIMATION_INTERVAL_KEY, config.test_animation_interval_sec, MAX_INTERVAL_SEC);
             config.test_animation_interval_sec = (config.test_animation_interval_sec < 0) ? 0 : MAX_INTERVAL_SEC;
+            ret = 2;
         }
         if (config.animation_speed_ms < 0 || config.animation_speed_ms > MAX_INTERVAL_SEC*1000) {
             BONGOCAT_LOG_WARNING("%s %d out of range [0-%dms], clamping",
                                  ANIMATION_SPEED_KEY, config.test_animation_interval_sec, MAX_INTERVAL_SEC*1000);
             config.animation_speed_ms = (config.animation_speed_ms < 0) ? 0 : MAX_INTERVAL_SEC*1000;
+            ret = 3;
         }
+        return ret;
     }
 
-    static void config_validate_kpm(config_t& config) {
-        config_clamp_int(config.happy_kpm, MIN_KPM, MAX_KPM, HAPPY_KPM_KEY);
+    static int32_t config_validate_kpm(config_t& config) {
+        return config_clamp_int(config.happy_kpm, MIN_KPM, MAX_KPM, HAPPY_KPM_KEY);
     }
 
-    static void config_validate_appearance(config_t& config) {
+    static int32_t config_validate_appearance(config_t& config) {
         using namespace assets;
+        int32_t ret{0};
         // Validate opacity
-        config_clamp_int(config.overlay_opacity, 0, 255, OVERLAY_OPACITY_KEY);
+        ret |= config_clamp_int(config.overlay_opacity, 0, 255, OVERLAY_OPACITY_KEY);
 
         switch (config.animation_sprite_sheet_layout) {
             case config_animation_sprite_sheet_layout_t::None:
                 BONGOCAT_LOG_WARNING("Cant determine sprite sheet layout");
+                ret = 2;
                 break;
             case config_animation_sprite_sheet_layout_t::Bongocat:
                 if constexpr (features::EnableBongocatEmbeddedAssets) {
@@ -161,6 +171,7 @@ namespace bongocat::config {
                         BONGOCAT_LOG_WARNING("%s %d out of range [0-%d], resetting to 0",
                                              ANIMATION_INDEX_KEY, config.animation_index, assets::BONGOCAT_ANIMATIONS_COUNT - 1);
                         config.animation_index = 0;
+                        ret = 3;
                     }
                     // Validate idle frame
                     assert(animation::BONGOCAT_NUM_FRAMES <= INT_MAX);
@@ -168,6 +179,7 @@ namespace bongocat::config {
                         BONGOCAT_LOG_WARNING("%s %d out of range [0-%d], resetting to 0",
                                              IDLE_FRAME_KEY, config.idle_frame, animation::BONGOCAT_NUM_FRAMES - 1);
                         config.idle_frame = 0;
+                        ret = 4;
                     }
                 }
                 break;
@@ -179,6 +191,7 @@ namespace bongocat::config {
                         BONGOCAT_LOG_WARNING("%s %d out of range [0-%d], resetting to 0",
                                              ANIMATION_INDEX_KEY, config.animation_index, assets::DM_ANIMATIONS_COUNT - 1);
                         config.animation_index = 0;
+                        ret = 5;
                     }
                     // Validate idle frame
                     assert(animation::MAX_DIGIMON_FRAMES <= INT_MAX);
@@ -186,6 +199,7 @@ namespace bongocat::config {
                         BONGOCAT_LOG_WARNING("%s %d out of range [0-%d], resetting to 0",
                                              IDLE_FRAME_KEY, config.idle_frame, animation::MAX_DIGIMON_FRAMES - 1);
                         config.idle_frame = 0;
+                        ret = 6;
                     }
                 }
                 break;
@@ -197,6 +211,7 @@ namespace bongocat::config {
                         BONGOCAT_LOG_WARNING("%s %d out of range [0-%d], resetting to 0",
                                              ANIMATION_INDEX_KEY, config.animation_index, assets::MS_AGENTS_ANIMATIONS_COUNT - 1);
                         config.animation_index = 0;
+                        ret = 7;
                     }
                     // Validate idle frame
                     assert(assets::MAX_SPRITE_SHEET_COL_FRAMES <= INT_MAX);
@@ -204,33 +219,42 @@ namespace bongocat::config {
                         BONGOCAT_LOG_WARNING("%s %d out of range [0-%d], resetting to 0",
                                              IDLE_FRAME_KEY, config.idle_frame, assets::MAX_SPRITE_SHEET_COL_FRAMES - 1);
                         config.idle_frame = 0;
+                        ret = 8;
                     }
                 }
                 break;
         }
+        return ret;
     }
 
-    static void config_validate_enums(config_t& config) {
+    static int32_t config_validate_enums(config_t& config) {
+        int32_t ret{0};
         // Validate layer
         if (config.layer != layer_type_t::LAYER_TOP && config.layer != layer_type_t::LAYER_OVERLAY) {
             BONGOCAT_LOG_WARNING("Invalid layer %d, resetting to top", config.layer);
             config.layer = layer_type_t::LAYER_TOP;
+            ret = 1;
         }
 
         // Validate overlay_position
         if (config.overlay_position != overlay_position_t::POSITION_TOP && config.overlay_position != overlay_position_t::POSITION_BOTTOM) {
             BONGOCAT_LOG_WARNING("Invalid %s %d, resetting to top", OVERLAY_OPACITY_KEY, config.overlay_position);
             config.overlay_position = overlay_position_t::POSITION_TOP;
+            ret = 2;
         }
 
         // Validate cat_align
         if (config.cat_align != align_type_t::ALIGN_CENTER && config.cat_align != align_type_t::ALIGN_LEFT && config.cat_align != align_type_t::ALIGN_RIGHT) {
             BONGOCAT_LOG_WARNING("Invalid %s %d, resetting to center", CAT_ALIGN_KEY, config.cat_align);
             config.cat_align = align_type_t::ALIGN_CENTER;
+            ret = 3;
         }
+
+        return ret;
     }
 
-    static void config_validate_time(config_t& config) {
+    static int32_t config_validate_time(config_t& config) {
+        int32_t ret{0};
         if (config.enable_scheduled_sleep) {
             const int begin_minutes = config.sleep_begin.hour * 60 + config.sleep_begin.min;
             const int end_minutes = config.sleep_end.hour * 60 + config.sleep_end.min;
@@ -243,11 +267,14 @@ namespace bongocat::config {
                 //config.sleep_begin.min = 0;
                 //config.sleep_end.hour = 0;
                 //config.sleep_end.min = 0;
+                ret = 1;
             }
         }
+        return ret;
     }
 
     static bongocat_error_t config_validate(config_t& config) {
+        int32_t ret{0};
         // Normalize boolean values
         config.enable_debug = config.enable_debug ? 1 : 0;
         config.invert_color = config.invert_color ? 1 : 0;
@@ -258,13 +285,19 @@ namespace bongocat::config {
         config.randomize_index = config.randomize_index ? 1 : 0;
         config.enable_antialiasing = config.enable_antialiasing ? 1 : 0;
 
-        config_validate_dimensions(config);
-        config_validate_timing(config);
-        config_validate_appearance(config);
-        config_validate_enums(config);
-        config_validate_time(config);
-        config_validate_kpm(config);
+        ret |= config_validate_dimensions(config);
+        ret |= config_validate_timing(config);
+        ret |= config_validate_appearance(config);
+        ret |= config_validate_enums(config);
+        ret |= config_validate_time(config);
+        ret |= config_validate_kpm(config);
 
+        if (config.strict) {
+            if (ret != 0) {
+                BONGOCAT_LOG_ERROR("Failed to load configuration in struct mode: %x", ret);
+                return bongocat_error_t::BONGOCAT_ERROR_INVALID_PARAM;
+            }
+        }
         return bongocat_error_t::BONGOCAT_SUCCESS;
     }
 
@@ -282,7 +315,7 @@ namespace bongocat::config {
         assert(MAX_INPUT_DEVICES <= INT_MAX);
         if (old_num_keyboard_devices >= static_cast<int>(MAX_INPUT_DEVICES)) {
             BONGOCAT_LOG_WARNING("Can not add more devices from config, max. reach: %d", MAX_INPUT_DEVICES);
-            return bongocat_error_t::BONGOCAT_SUCCESS;
+            return config.strict ? bongocat_error_t::BONGOCAT_ERROR_INVALID_PARAM : bongocat_error_t::BONGOCAT_SUCCESS;
         }
         const int new_num_keyboard_devices = old_num_keyboard_devices + 1;
         assert(new_num_keyboard_devices >= 0);
@@ -569,6 +602,9 @@ namespace bongocat::config {
             if (config.animation_index < 0 || config.animation_sprite_sheet_layout == config_animation_sprite_sheet_layout_t::None) {
                 if (config.animation_index >= 0 && config.animation_sprite_sheet_layout == config_animation_sprite_sheet_layout_t::None) {
                     BONGOCAT_LOG_WARNING("animation_index is set, but not animation_type (unknown type for index=%i and value='%s')", config.animation_index, value);
+                    if (config.strict) {
+                        return bongocat_error_t::BONGOCAT_ERROR_INVALID_PARAM;
+                    }
                 }
                 BONGOCAT_LOG_WARNING("Invalid %s '%s', using '%s'", ANIMATION_NAME_KEY, value, BONGOCAT_NAME);
                 config.animation_index = BONGOCAT_ANIM_INDEX;
@@ -710,6 +746,7 @@ namespace bongocat::config {
         cfg.input_fps = 0;          // when 0 fallback to fps
         cfg.randomize_index = 0;
         cfg.keep_old_animation_index = 0;
+        cfg.strict = 0;
 
         config = bongocat::move(cfg);
     }
@@ -772,8 +809,11 @@ namespace bongocat::config {
             if (ret.output_name) ::free(ret.output_name);
             ret.output_name = strdup(overwrite_parameters.output_name);
         }
-        if (overwrite_parameters.random_index >= 0) {
-            ret.randomize_index = overwrite_parameters.random_index ? 1 : 0;
+        if (overwrite_parameters.randomize_index >= 0) {
+            ret.randomize_index = overwrite_parameters.randomize_index ? 1 : 0;
+        }
+        if (overwrite_parameters.strict >= 0) {
+            ret.strict = overwrite_parameters.strict ? 1 : 0;
         }
         if (ret.input_fps <= 0) {
             ret.input_fps = ret.fps;